requirements.txt
langchain-huggingface
python-dotenv
langchain
tensorflow
streamlit